Add webhook notifier support
We need a way to alert a server when a feed changes.
It isn't really feasible to handle, say, 100 requests per second
just to return a 304 for everything, if we can avoid it.
I mean, it's possible, just not ideal.

Instead, configure a WEBHOOK_URL,
and when rows are modified, they'll get a webhook dispatched
containing the modified URLs.
The caller can then make HTTP requests to get the new data.