Store feeds in S3/R2, not the DB
DB storage can be huge, at 5mb per feed.
For 30k feeds, we're using 150+GB.

We do not need feeds in Postgres for this since we aren't
really using relational patterns, just identifying a file.

Instead, store the feeds in S3 (or R2 or anything else compatible).

Since this is just a caching server, it doesn't seem necessary
to build a migration tool; instead, the cache will be empty
when this deploys. You must drop the contents table yourself,
since we don't want to delete it in case of rollbacks.